+++ test/flushbar_test.dart
@@ -1,6 +1,7 @@
 import 'dart:ui';
 
 import 'package:flushbar/flushbar.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -28,6 +29,8 @@ void main() {
     expect(flushbar.progressIndicatorController, null);
     expect(flushbar.progressIndicatorBackgroundColor, null);
     expect(flushbar.progressIndicatorValueColor, null);
+    expect(flushbar.overlayBlur, 0.0);
+    expect(flushbar.overlayColor, Colors.transparent);
     expect(flushbar.isShowing(), false);
     expect(flushbar.isDismissed(), false);
     expect(await flushbar.dismiss(), null);